// src/app/models/search-phone-calls.models.ts
export class SearchPhoneCalls {
    due_date_from: Date;
    due_date_to: Date;
    phone_call_title: string;
    reason_for_call: string;
    scheduled_by: number;
    assigned_to_user: number;
    assigned_to_role: number;
    call_result: string;
    call_status: number;
    project_number: number;
    customer_first_name: string;
    customer_last_name: string;
    estimator: number;
    crew: number;
}